requirements.txt
jpype1==1.5.2
packaging==24.2
